fix decorrelate join type choose
